Fixes for SSL support
Finally got a good compile from libwebsockets. CMAKE was defaulting
solution to x64 despite projects set to x86. Once that was compiled, I
put the new lib in JSMPEG and enabled SSL. The server runs but the
browser gives a message saying it's not secure. Also, the WS connection
does not work. Thinking I need to redo the code that sends the message
to align with the new WS librar version. Not sure yet.